moved profileViewer to profile directory
