+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
         <activity
-            android:name=".ProfileViewer"
+            android:name=".ui.profile.ProfileViewer"
             android:label="@string/title_activity_profile_viewer"
             android:parentActivityName=".ui.groupFinder.StudyGroupFinder"
             android:windowSoftInputMode="stateHidden|adjustResize">
