+++ app/src/main/java/study_dev/testbed/studyhopper/ui/profile/ProfileViewer.java
@@ -1,4 +1,4 @@
-package study_dev.testbed.studyhopper;
+package study_dev.testbed.studyhopper.ui.profile;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -21,12 +21,11 @@ import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.ArrayList;
 
+import study_dev.testbed.studyhopper.R;
 import study_dev.testbed.studyhopper.models.Profile;
 import study_dev.testbed.studyhopper.models.StudentClass;
 import study_dev.testbed.studyhopper.ui.dashboard.Dashboard;
 import study_dev.testbed.studyhopper.ui.groupFinder.StudyGroupFinder;
-import study_dev.testbed.studyhopper.ui.profile.ClassesListAdapter;
-import study_dev.testbed.studyhopper.ui.profile.classListItem;
 
 public class ProfileViewer extends AppCompatActivity {
     private DocumentReference userRef;
